requirements.txt
# Odysseia-Similu 音乐机器人依赖包
# Core Discord bot functionality
discord.py[voice]>=2.0.0

# Configuration management
pyyaml>=6.0

# HTTP requests and async operations
requests>=2.28.0
aiohttp>=3.8.0

# Audio processing
python-ffmpeg>=1.0.0

# YouTube video processing
pytubefix>=6.0.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.18.0